+++ test/generator_test.rb
@@ -90,7 +90,7 @@ class GeneratorTest < Minitest::Test
     <title>{{ page.title }}</title>
     {% assign css_files = files | where: "extname", ".css" %}
     {% for file in css_files %}
-      <link rel="stylesheet" href="{{ file.url | relative_to: page.url }}">
+      <link rel="stylesheet" href="{{ file.path | relative }}">
     {% endfor %}
   </head>
   <body>