+++ test/source_test.rb
@@ -7,9 +7,7 @@ class SourceTest < Minitest::Test
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("index.html"))
 
-      assert_equal "index.html", source.relative_path.to_s
-      assert_equal "index.html", source.dest_path.basename.to_s
-      assert_equal ".", source.url.to_s
+      assert_equal "index.html", source.dest_path.to_s
       assert_equal "Home", source.title
       assert_nil source.layout
       refute source.dynamic?
@@ -20,9 +18,7 @@ class SourceTest < Minitest::Test
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("foo", "bar", "index.html"))
 
-      assert_equal "foo/bar/index.html", source.relative_path.to_s
-      assert_equal "index.html", source.dest_path.basename.to_s
-      assert_equal "foo/bar", source.url.to_s
+      assert_equal "foo/bar/index.html", source.dest_path.to_s
       assert_equal "Bar", source.title
       assert_nil source.layout
       refute source.dynamic?
@@ -33,9 +29,7 @@ class SourceTest < Minitest::Test
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("foo.html"))
 
-      assert_equal "foo.html", source.relative_path.to_s
-      assert_equal "foo.html", source.dest_path.basename.to_s
-      assert_equal "foo.html", source.url.to_s
+      assert_equal "foo.html", source.dest_path.to_s
       assert_equal "Foo", source.title
       assert_nil source.layout
       refute source.dynamic?
@@ -46,9 +40,7 @@ class SourceTest < Minitest::Test
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("foo/bar/foo.html"))
 
-      assert_equal "foo/bar/foo.html", source.relative_path.to_s
-      assert_equal "foo.html", source.dest_path.basename.to_s
-      assert_equal "foo/bar/foo.html", source.url.to_s
+      assert_equal "foo/bar/foo.html", source.dest_path.to_s
       assert_equal "Foo", source.title
       assert_nil source.layout
       refute source.dynamic?
@@ -60,9 +52,7 @@ class SourceTest < Minitest::Test
       File.write(config.layout_dir.join("page.html.liquid"), "")
       source = Yass::Source.new(config, config.src_dir.join("foo/bar/foo.page.html"))
 
-      assert_equal "foo/bar/foo.page.html", source.relative_path.to_s
-      assert_equal "foo.html", source.dest_path.basename.to_s
-      assert_equal "foo/bar/foo.html", source.url.to_s
+      assert_equal "foo/bar/foo.html", source.dest_path.to_s
       assert_equal "Foo", source.title
       refute_nil source.layout
       assert_equal config.layout_cache["page.html"], source.layout
@@ -75,9 +65,7 @@ class SourceTest < Minitest::Test
       File.write(config.layout_dir.join("post.html.liquid"), "")
       source = Yass::Source.new(config, config.src_dir.join("2025/01/01/my-post.post.md"))
 
-      assert_equal "2025/01/01/my-post.post.md", source.relative_path.to_s
-      assert_equal "my-post.html", source.dest_path.basename.to_s
-      assert_equal "2025/01/01/my-post.html", source.url.to_s
+      assert_equal "2025/01/01/my-post.html", source.dest_path.to_s
       assert_equal "My Post", source.title
       refute_nil source.layout
       assert_equal config.layout_cache["post.html"], source.layout
@@ -88,9 +76,7 @@ class SourceTest < Minitest::Test
   def test_md_files
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("2025/01/01/my-post.md"))
-      assert_equal "2025/01/01/my-post.md", source.relative_path.to_s
-      assert_equal "my-post.html", source.dest_path.basename.to_s
-      assert_equal "2025/01/01/my-post.html", source.url.to_s
+      assert_equal "2025/01/01/my-post.html", source.dest_path.to_s
       assert_equal "My Post", source.title
       assert_nil source.layout
       assert source.dynamic?
@@ -102,9 +88,7 @@ class SourceTest < Minitest::Test
       File.write(config.layout_dir.join("post.html.liquid"), "")
       source = Yass::Source.new(config, config.src_dir.join("2025/01/01/my-post.post.html.liquid"))
 
-      assert_equal "2025/01/01/my-post.post.html.liquid", source.relative_path.to_s
-      assert_equal "my-post.html", source.dest_path.basename.to_s
-      assert_equal "2025/01/01/my-post.html", source.url.to_s
+      assert_equal "2025/01/01/my-post.html", source.dest_path.to_s
       assert_equal "My Post", source.title
       refute_nil source.layout
       assert_equal config.layout_cache["post.html"], source.layout
@@ -137,9 +121,7 @@ class SourceTest < Minitest::Test
   def test_liquid_files
     with_config do |config|
       source = Yass::Source.new(config, config.src_dir.join("2025/01/01/my-post.html.liquid"))
-      assert_equal "2025/01/01/my-post.html.liquid", source.relative_path.to_s
-      assert_equal "my-post.html", source.dest_path.basename.to_s
-      assert_equal "2025/01/01/my-post.html", source.url.to_s
+      assert_equal "2025/01/01/my-post.html", source.dest_path.to_s
       assert_equal "My Post", source.title
       assert_nil source.layout
       assert source.dynamic?